+++ R/calc_extremes.R
@@ -1,6 +1,6 @@
 #' Calculate the time spent near the extremes
 #'
-#' This function calculates the time the subject spent near the set minimum and maximum temperature limits during the shuttle-box trial
+#' This function calculates the percentage of observations spent near the set minimum and maximum temperature limits during the shuttle-box trial
 #'
 #' @param data An organised shuttle-box dataframe with corrected core body temperature
 #' @param threshold Definition of the extreme temperature range. Default is 20% of the temperature range: 0.2*(max(data$max_T)-max(data$min_T))
@@ -8,7 +8,7 @@
 #' @param exclude_start_minutes Exclusion of time from the start of the trial onwards, in minutes. Default is 0
 #' @param exclude_end_minutes Exclusion of time from the end of the trial backwards, in minutes. Default is 0
 #' @param print_results Print the results, default is TRUE
-#' @return the time spent near each extreme
+#' @return A two-element vector giving the percentage of observations near the lower and upper extremes
 #' @export
 
 calc_extremes <- function(data, threshold = 0.2*(max(data$max_T)-max(data$min_T)), exclude_start_minutes = 0, exclude_end_minutes = 0, exclude_acclimation = F, print_results = T) {